vc: long -> int32

+++ src/cmd/vc/cgen.c
@@ -7,7 +7,7 @@ cgen(Node *n, Node *nn)
 	Prog *p1;
 	Node nod, nod1, nod2, nod3, nod4;
 	int o;
-	long v, curs;
+	int32 v, curs;
 
 	if(debug['g']) {
 		prtree(nn, "cgen lhs");
@@ -410,7 +410,7 @@ cgen(Node *n, Node *nn)
 				diag(n, "DOT and no offset");
 				break;
 			}
-			nod.xoffset += (long)r->vconst;
+			nod.xoffset += (int32)r->vconst;
 			nod.type = n->type;
 			cgen(&nod, nn);
 		}
@@ -526,7 +526,7 @@ void
 reglcgen(Node *t, Node *n, Node *nn)
 {
 	Node *r;
-	long v;
+	int32 v;
 
 	regialloc(t, n, nn);
 	if(n->op == OIND) {
@@ -615,7 +615,7 @@ boolgen(Node *n, int true, Node *nn)
 	int o;
 	Prog *p1, *p2;
 	Node *l, *r, nod, nod1;
-	long curs;
+	int32 curs;
 
 	if(debug['g']) {
 		prtree(nn, "boolgen lhs");
@@ -858,12 +858,12 @@ boolgen(Node *n, int true, Node *nn)
 }
 
 void
-sugen(Node *n, Node *nn, long w)
+sugen(Node *n, Node *nn, int32 w)
 {
 	Prog *p1;
 	Node nod0, nod1, nod2, nod3, nod4, *l, *r;
 	Type *t;
-	long pc1;
+	int32 pc1;
 	int i, m, c;
 
 	if(n == Z || n->type == T)
@@ -923,7 +923,7 @@ sugen(Node *n, Node *nn, long w)
 				diag(n, "DOT and no offset");
 				break;
 			}
-			nod1.xoffset += (long)r->vconst;
+			nod1.xoffset += (int32)r->vconst;
 			nod1.type = n->type;
 			sugen(&nod1, nn, w);
 		}
